# docs/mobile-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIN Mobile API Contract",
    "version": "2026-07-24",
    "description": "Flutter-facing contract for the mobile endpoints currently implemented in the Laravel API."
  },
  "servers": [
    {
      "url": "https://api.example.com/api/v1",
      "description": "Production API base URL placeholder"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local Laravel API"
    }
  ],
  "security": [
    {
      "sanctumBearer": []
    }
  ],
  "paths": {
...
    "/auth/verify-email": {
      "post": {
        "operationId": "verifyEmail",
        "tags": [
          "Auth"
        ],
        "security": [],
        "summary": "Verify account email and return a mobile session.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verified session.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthSession"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "resendVerification",
        "tags": [
          "Auth"
        ],
        "security": [],
        "summary": "Resend an email verification code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification resend accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AcceptedDeliveryResult"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "forgotPassword",
        "tags": [
          "Auth"
        ],
        "security": [],
        "summary": "Request a password reset code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Password reset request accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AcceptedDeliveryResult"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "tags": [
          "Auth"
        ],
        "security": [],
        "summary": "Reset password using the emailed code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PasswordResetResult"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sanctumBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum"
      }
    },
...
    "responses": {
      "ApiError": {
        "description": "Standard API error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "data",
          "meta",
          "requestId"
        ],
        "properties": {
          "data": {},
          "meta": {
            "type": "object"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "catalog",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "catalog": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ErrorCatalogEntry"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
      "ErrorCatalogEntry": {
        "type": "object",
        "required": [
          "category",
          "retryable",
          "messageEn",
          "messageAr"
        ],
        "properties": {
          "category": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "messageEn": {
            "type": "string"
          },
          "messageAr": {
            "type": "string"
          }
        }
      },
...
      "EmailRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "EmailCodeRequest": {
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        }
      },
      "ResetPasswordRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EmailCodeRequest"
          },
          {
            "type": "object",
            "required": [
              "password",
              "password_confirmation"
            ],
            "properties": {
              "password": {
                "type": "string",
                "minLength": 10
              },
              "password_confirmation": {
                "type": "string",
                "minLength": 10
              }
            }
          }
        ]
      },
...
      "AcceptedDeliveryResult": {
        "type": "object",
        "required": [
          "accepted",
          "delivery"
        ],
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "delivery": {
            "type": "string",
            "enum": [
              "email"
            ]
          },
          "debugCode": {
            "type": "string"
          }
        }
      },
      "PasswordResetResult": {
        "type": "object",
        "required": [
          "passwordChanged"
        ],
        "properties": {
          "passwordChanged": {
            "type": "boolean"
          }
        }
      },
      "AuthSession": {
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "type": "object"
          }
        }
      },
...
    }
  }
}